Fixes the constant re-downloading of image files.
In the previous slider, images would be downloaded each time dynamically
because the background-url property would constantly be reset.  In order
to avoid this and reduce network traffic, I've created the slide
immediately and pushed it into the data store so that we can just access
it.  This resolves Todo #14.